requirements.txt
transformers
torch
scipy
mysql-connector-python